\name{guess_datetime_format}
\alias{guess_datetime_format}
\title{Guess strptime format of a date time string}
\usage{
  guess_datetime_format(string, default = "mdy")
}
\arguments{
  \item{string}{date time string whose format is to be
  guessed}
}
\value{
  a string consisting of strptime tokens
}
\description{
  Guess strptime format of a date time string
}

