\name{split_datetime_string}
\alias{split_datetime_string}
\title{Splits a date time string using regex operators from the operators package}
\usage{
  split_datetime_string(string)
}
\description{
  Leads to extremely compact code
}
\examples{
split_datetime_string("20 Jan 10:30 PM on Sunday")
split_datetime_string("20 Jan 10:30 PM")
split_datetime_string("10:30:14 on Sunday")
}

